refactor: remove io_utils compatibility layer
