+++ src/tnfr/config.py
@@ -2,7 +2,7 @@
 from __future__ import annotations
 from typing import Any, Dict
 from pathlib import Path
-from .io_utils import read_structured_file
+from .helpers import read_structured_file
 
 from .constants import inject_defaults
