--- src/tnfr/io_utils.py
@@ -1,7 +1,0 @@
-"""Compatibility layer for structured file helpers."""
-from __future__ import annotations
-
-from .helpers import read_structured_file, ensure_parent
-
-__all__ = ["read_structured_file", "ensure_parent"]
-
